Publish migration instead of running from package
Running the migration from the package does not allow customisation
of the database table to fit the application.
Also, it does not allow the user to determine when the migration is run
in relation to migrations.

If the migration is published, the user can then modify both the
timestamp to dictate when it is run and the column types to suit
the needs of the app.

E.g. changing ->morphs() to ->uuidMorphs for users using uuid primary
key fields.